+++ opentelemetry-sdk/src/logs/log_processor.rs
@@ -125,6 +125,11 @@ pub(crate) mod tests {
                 Key::from_static_str("processed_by"),
                 AnyValue::String("FirstProcessor".into()),
             );
+            record.add_attribute(
+                Key::from_static_str("key1"),
+                AnyValue::String("val1".into()),
+            );
+
             // update body
             record.body = Some("Updated by FirstProcessor".into());
 
@@ -154,6 +159,51 @@ pub(crate) mod tests {
                 &Key::from_static_str("processed_by"),
                 &AnyValue::String("FirstProcessor".into())
             ));
+            record.update_attribute(
+                &Key::from_static_str("processed_by"),
+                AnyValue::from("SecondProcessor"),
+            );
+            let _ = record.remove_attribute(&Key::from_static_str("key1"));
+            assert!(
+                record.body.clone().unwrap()
+                    == AnyValue::String("Updated by FirstProcessor".into())
+            );
+            self.logs
+                .lock()
+                .unwrap()
+                .push((record.clone(), instrumentation.clone()));
+        }
+
+        fn force_flush(&self) -> OTelSdkResult {
+            Ok(())
+        }
+
+        fn shutdown(&self) -> OTelSdkResult {
+            Ok(())
+        }
+    }
+
+    #[derive(Debug)]
+    struct ThirdProcessor {
+        pub(crate) logs: Arc<Mutex<Vec<(SdkLogRecord, InstrumentationScope)>>>,
+    }
+
+    impl LogProcessor for ThirdProcessor {
+        fn emit(&self, record: &mut SdkLogRecord, instrumentation: &InstrumentationScope) {
+            assert!(record.attributes_contains(
+                &Key::from_static_str("processed_by"),
+                &AnyValue::String("SecondProcessor".into())
+            ));
+            record.update_attribute(
+                &Key::from_static_str("processed_by"),
+                AnyValue::from("ThirdProcessor"),
+            );
+            assert!(!record.attributes_contains(
+                &Key::from_static_str("key1"),
+                &AnyValue::String("value1".into())
+            ));
+
+            let _ = record.remove_attribute(&Key::from_static_str("key1"));
             assert!(
                 record.body.clone().unwrap()
                     == AnyValue::String("Updated by FirstProcessor".into())
@@ -177,6 +227,7 @@ pub(crate) mod tests {
     fn test_log_data_modification_by_multiple_processors() {
         let first_processor_logs = Arc::new(Mutex::new(Vec::new()));
         let second_processor_logs = Arc::new(Mutex::new(Vec::new()));
+        let third_processor_logs = Arc::new(Mutex::new(Vec::new()));
 
         let first_processor = FirstProcessor {
             logs: Arc::clone(&first_processor_logs),
@@ -184,10 +235,14 @@ pub(crate) mod tests {
         let second_processor = SecondProcessor {
             logs: Arc::clone(&second_processor_logs),
         };
+        let third_processor = ThirdProcessor {
+            logs: Arc::clone(&third_processor_logs),
+        };
 
         let logger_provider = SdkLoggerProvider::builder()
             .with_log_processor(first_processor)
             .with_log_processor(second_processor)
+            .with_log_processor(third_processor)
             .build();
 
         let logger = logger_provider.logger("test-logger");
@@ -198,9 +253,11 @@ pub(crate) mod tests {
 
         assert_eq!(first_processor_logs.lock().unwrap().len(), 1);
         assert_eq!(second_processor_logs.lock().unwrap().len(), 1);
+        assert_eq!(third_processor_logs.lock().unwrap().len(), 1);
 
         let first_log = &first_processor_logs.lock().unwrap()[0];
         let second_log = &second_processor_logs.lock().unwrap()[0];
+        let third_log = &third_processor_logs.lock().unwrap()[0];
 
         assert!(first_log.0.attributes_contains(
             &Key::from_static_str("processed_by"),
@@ -208,7 +265,11 @@ pub(crate) mod tests {
         ));
         assert!(second_log.0.attributes_contains(
             &Key::from_static_str("processed_by"),
-            &AnyValue::String("FirstProcessor".into())
+            &AnyValue::String("SecondProcessor".into())
+        ));
+        assert!(third_log.0.attributes_contains(
+            &Key::from_static_str("processed_by"),
+            &AnyValue::String("ThirdProcessor".into())
         ));
 
         assert!(
@@ -217,6 +278,10 @@ pub(crate) mod tests {
         );
         assert!(
             second_log.0.body.clone().unwrap()
+                == AnyValue::String("Updated by FirstProcessor".into())
+        );
+        assert!(
+            third_log.0.body.clone().unwrap()
                 == AnyValue::String("Updated by FirstProcessor".into())
         );
     }